Rename to `@aegisjsproject/core`
