+++ package.json
@@ -1,6 +1,6 @@
 {
-  "name": "@shgysk8zer0/aegis",
-  "version": "0.0.8",
+  "name": "@aegisjsproject/core",
+  "version": "0.1.0",
   "description": "A fast, secure, modern, light-weight, and simple JS library for creating web components and more!",
   "keywords": [
     "aegis",
@@ -17,13 +17,13 @@
     "component-library"
   ],
   "type": "module",
-  "main": "./aegis.cjs",
-  "module": "./aegis.js",
-  "unpkg": "./aegis.js",
+  "main": "./core.cjs",
+  "module": "./core.js",
+  "unpkg": "./core.js",
   "exports": {
     ".": {
-      "import": "./aegis.js",
-      "require": "./aegis.cjs"
+      "import": "./core.js",
+      "require": "./core.cjs"
     },
     "./*.js": {
       "import": "./*.js",
@@ -71,7 +71,7 @@
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/shgysk8zer0/aegis.git"
+    "url": "git+https://github.com/AegisJSProject/core.git"
   },
   "author": "Chris Zuber <admin@kernvalley.us>",
   "license": "MIT",
@@ -86,9 +86,9 @@
     }
   ],
   "bugs": {
-    "url": "https://github.com/shgysk8zer0/aegis/issues"
+    "url": "https://github.com/AegisJSProject/core/issues"
   },
-  "homepage": "https://github.com/shgysk8zer0/aegis#readme",
+  "homepage": "https://github.com/AegisJSProject/core#readme",
   "devDependencies": {
     "@shgysk8zer0/aegis-component": "^0.0.3",
     "@shgysk8zer0/aegis-styles": "^0.0.3",
