+++ test/index.html
@@ -6,22 +6,26 @@
 		<meta name="color-scheme" content="light dark" />
 		<meta http-equiv="Content-Security-Policy"
 			content="default-src 'none';
-			script-src 'self' https://unpkg.com/@shgysk8zer0/ 'sha384-sEZHY5s3EVOtSOk2on2ohawJgDJCF2syHikVzJwj47nzEf5gsziUWZ2QMcmzli7C';
+			script-src 'self' https://unpkg.com/@shgysk8zer0/ https://unpkg.com/@aegisjsproject/ 'sha384-z00C+r/e+I1mYMOID0+ivQIyrVhwtoXQZ75ohGBL1hgHzznjQSR4Rg0l4jQp22H4';
 			style-src 'self' blob: data:;
-			connect-src https://icanhazdadjoke.com https://unpkg.com/@shgysk8zer0/;
+			connect-src https://icanhazdadjoke.com https://unpkg.com/@shgysk8zer0/ https://unpkg.com/@kernvalley/;
 			trusted-types empty#html empty#script sanitizer-raw#html default;
 			require-trusted-types-for 'script';" />
 		<title>Aegis Test</title>
-		<script type="importmap" integrity="sha384-sEZHY5s3EVOtSOk2on2ohawJgDJCF2syHikVzJwj47nzEf5gsziUWZ2QMcmzli7C">
+		<script type="importmap" integrity="sha384-z00C+r/e+I1mYMOID0+ivQIyrVhwtoXQZ75ohGBL1hgHzznjQSR4Rg0l4jQp22H4">
 			{
 				"imports": {
-					"@shgysk8zer0/aegis": "../aegis.js",
+					"@aegisjsproject/core": "../core.js",
+					"@shgysk8zer0/aegis": "../core.js",
+					"@aegisjsproject/core/": "../",
 					"@shgysk8zer0/aegis/": "../",
+					"@shgysk8zer0/aegis/aegis.js": "../core.js",
 					"@shgysk8zer0/aegis-styles": "../node_modules/@shgysk8zer0/aegis-styles/styles.js",
 					"@shgysk8zer0/aegis-styles/": "../node_modules/@shgysk8zer0/aegis-styles/",
 					"@shgysk8zer0/aegis-component": "../node_modules/@shgysk8zer0/aegis-component/component.js",
 					"@shgysk8zer0/aegis-component/": "../node_modules/@shgyk8zer0/aegis-component/",
-					"@shgysk8zer0/": "../node_modules/@shgysk8zer0/"
+					"@shgysk8zer0/": "https://unpkg.com/@shgysk8zer0/",
+					"@kernvalley/": "https://unpkg.com/@kernvalley/"
 				}
 			}
 		</script>
